Get posts from all pages and add option to stop downloads at a post ID
Previous version of the script was only downloading posts that were in the first
page of the forum, now we can use pagination to see all posts. Also added a new
environment variable to stop at a minimum post ID so that people can download
only the posts until a certain ID and don't have to download everything if they
don't want to.